Practice separating and including files, did so with the helper functions used in views.py
