+++ learning_assignments/book_review_3/apps/book_review/views.py
@@ -2,17 +2,8 @@ from django.shortcuts import render, redirect
 from django.core.urlresolvers import reverse
 from django.contrib import messages
 from .models import User, Book, Author, Review
+from .functions import *
 
-def check_id(request):
-    if not 'id' in request.session:
-        return redirect(reverse('br:index'))
-
-def check_method(request):
-    if request.method != 'Post':
-        return redirect(reverse('br:index'))
-def flash_messages(request, errors):
-    for error in errors:
-        messages.info(request, error)
